implemented service that can count amount salary with all info that needs CRM view

+++ src/main/java/com/example/OvertimeTracker/service/salaryCounter/perhourCounter/SalaryPerHourService.java
@@ -1,9 +1,10 @@
 package com.example.OvertimeTracker.service.salaryCounter.perhourCounter;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 public interface SalaryPerHourService {
 
-    BigInteger getPerHourValue(BigInteger baseSalary);
+    BigDecimal getPerHourValue(BigDecimal baseSalary);
 
 }
